fix: Corrected hooks for Atom submenus and fixed "main" misdirection
As it turns out I learned a lot from trial and error in this commit.
Unfortunately, Atom doesn't warn the user if one of it's plugins
contain a malformed package.json (outside of syntax errors). Which
means I was in the dark for a while working on getting this fixed.
If the "main" property has a value which points to a nonexistant
file, the plugin will appear to be installed and functioning as
normal in both settings view and the submenus, but without any
real functionality. When the broken stubs are called. The dysfunctional
package is automatically discarded from Atom's working pool and no
error is reported. This behavior is extremely unhelpful. But now
that I know about it I can (after navigating the hellfire which is
Atom contribution beurocracy) submit a bug report. If it's not fixed
I at least know of the problem's existance so I can use it as a
further diagnostic skill when working with Atom.